+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+- The token is also passed to the policies
 - CI for the graphql part
 - The last tenant selected would be displayed
 - Possibility do add custom icons (with some size limitaion) to your tenant